refactor: move NavLink into header folder
